# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Comprehensive REST API for managing game server infrastructure with panel integration, async job queue management, and real-time sync operations",
        "title": "NodeByte API",
        "termsOfService": "https://nodebyte.co.uk/legal/terms",
        "contact": {
            "name": "Contact Support",
            "url": "https://discord.gg/wN58bTzzpW"
        },
        "license": {
            "name": "AGPL 3.0",
            "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
        },
        "version": "0.2.0"
    },
    "host": "core.nodebyte.host",
    "basePath": "/",
    "paths": {
...
        "/api/v1/hytale/logs": {
            "get": {
                "description": "Retrieves audit logs for a specific Hytale account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hytale Logs"
                ],
                "summary": "Get Hytale Audit Logs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account UUID",
                        "name": "account_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Maximum number of logs (default: 100, max: 1000)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.GetHytaleLogsResponse"
                        }
                    },
                    "400": {
                        "description": "Missing account_id or invalid parameters",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Human-readable error message",
                    "type": "string",
                    "example": "account_id is required"
                },
                "success": {
                    "description": "Whether the request was successful",
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "types.GetHytaleLogsResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer",
                    "example": 50
                },
                "logs": {},
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Backend-to-backend API authentication",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}